requirements.txt
appnope==0.1.0
cycler==0.10.0
dask==0.13.0
decorator==4.0.11
ipython==5.2.2
ipython-genutils==0.1.0
#matlabengineforpython===R2016b
matplotlib==2.0.0
networkx==1.11
numpy==1.12.0
olefile==0.44
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.0.0
prompt-toolkit==1.0.13
ptyprocess==0.5.1
Pygments==2.2.0
pyparsing==2.1.10
python-dateutil==2.6.0
pytz==2016.10
scikit-image==0.12.3
scikit-learn==0.18.1
scipy==0.18.1
simplegeneric==0.8.1
six==1.10.0
toolz==0.8.2
traitlets==4.3.1
wcwidth==0.1.7
